requirements.txt
google-genai==1.27.0
pydantic==2.11.7
requests==2.32.4
typing-inspection==0.4.1
typing_extensions==4.14.1
streamlit
typing
